+++ src/insiphy/__init__.py
@@ -1,3 +1,3 @@
 """INSIPHY: intragenic synteny phylogenetics."""
 
-__version__ = "0.12.0"
+__version__ = "0.12.1"